+++ src/streams/inet/istream_tcp.c
@@ -33,7 +33,7 @@ static int32_t tk_istream_tcp_read(tk_istream_t* stream, uint8_t* buff, uint32_t
 
   ret = recv(istream_tcp->sock, buff, max_size, 0);
   if (ret <= 0) {
-    if (errno != EAGAIN) {
+    if (socket_is_last_io_ok()) {
       perror("recv");
       istream_tcp->is_broken = TRUE;
     }